Enhance cardinality handling in cost model constants
- Imported the cardinality package to utilize the Cardinality type for DefaultTableCardinality, improving type safety in cost estimation.
- Updated the EstimatePlanCost method to convert cardinality to int64 before setting it on plan nodes, ensuring consistent type usage across the cost model.

+++ pkg/optimizer/internal/cost_model/cost_model.go
@@ -71,7 +71,7 @@ func (cm *CostModel) EstimatePlanCost(planNode plan.PlanNode) float64 {
 		if err != nil {
 			card = DefaultTableCardinality
 		}
-		planNode.SetCardinality(card)
+		planNode.SetCardinality(int64(card))
 	}
 
 	var cost float64
